Test for channel keeper
